YG-187 💡 유저 프로필 수정 관련 함수 분리
